+++ R/anlzMWRoutlierall.R
@@ -50,7 +50,7 @@
 #' # create png output
 #' anlzMWRoutlierall(resdat, accdat, group = 'month', format = 'png', output_dir = tempdir())
 #' }
-anlzMWRoutlierall <- function(res = NULL, acc = NULL, fset = NULL, fig_height = 4, fig_width = 8, format = c('word' ,'png'), output_dir, output_file = NULL, type = c('box', 'jitterbox', 'jitter'), group, dtrng = NULL, repel = TRUE, outliers = FALSE, labsize = 3, fill = 'lightgrey', alpha = 0.8, width = 0.8, yscl = c('auto', 'log', 'linear'), ttlsize = 1.2, runchk = TRUE, warn = TRUE){
+anlzMWRoutlierall <- function(res = NULL, acc = NULL, fset = NULL, fig_height = 4, fig_width = 8, format = c('word' ,'png'), output_dir, output_file = NULL, type = c('box', 'jitterbox', 'jitter'), group, dtrng = NULL, repel = TRUE, outliers = FALSE, labsize = 3, fill = 'lightgrey', alpha = 0.8, width = 0.8, yscl = 'auto', ttlsize = 1.2, runchk = TRUE, warn = TRUE){
   
   utilMWRinputcheck(mget(ls()))
   
@@ -98,7 +98,9 @@ anlzMWRoutlierall <- function(res = NULL, acc = NULL, fset = NULL, fig_height = 
     
     if(is.null(output_file))
       output_file <- gsub('\\.Rmd$', '.docx', basename(outlierall))
-    file_loc <- list.files(path = output_dir, pattern = output_file, full.names = TRUE)
+    
+    output_file <- paste0(tools::file_path_sans_ext(output_file), '.docx')
+    file_loc <- list.files(path = output_dir, pattern = paste0('^', output_file), full.names = TRUE)
     msg <- paste("Word document created successfully! File located at", file_loc)
     message(msg)