+++ R/tabMWRwqx.R
@@ -121,7 +121,7 @@ tabMWRwqx <- function(res = NULL, acc = NULL, sit = NULL, wqx = NULL, fset = NUL
   
   ##
   # Results
-
+  
   # format parameter in accdat to wqx parameter
   accdat <- accdat %>% 
     dplyr::mutate(
@@ -195,8 +195,9 @@ tabMWRwqx <- function(res = NULL, acc = NULL, sit = NULL, wqx = NULL, fset = NUL
       `Result Value`, 
       `Result Unit`, 
       `Result Measure Qualifier`, 
-      `Result Comment`, 
-      `Quantitation Limit`
+      `Result Comment`,
+      `Quantitation Limit`,
+      `Activity ID User Supplied` = `Local Record ID`,
       ) %>% 
     dplyr::group_by(`Activity Start Date`, `Characteristic Name`, `Activity Type`) %>% 
     dplyr::mutate(
@@ -386,6 +387,7 @@ tabMWRwqx <- function(res = NULL, acc = NULL, sit = NULL, wqx = NULL, fset = NUL
       `Project ID`,
       `Monitoring Location ID`,
       `Activity ID`,
+      `Activity ID User Supplied`,
       `Activity Type`,
       `Activity Media Name`,
       `Activity Start Date`,
@@ -428,9 +430,10 @@ tabMWRwqx <- function(res = NULL, acc = NULL, sit = NULL, wqx = NULL, fset = NUL
   )
   
   # save
+  output_file <- paste0(tools::file_path_sans_ext(output_file), '.xlsx')
   writexl::write_xlsx(out, path = file.path(output_dir, output_file))
-  
-  file_loc <- list.files(path = output_dir, pattern = output_file, full.names = TRUE)
+
+  file_loc <- list.files(path = output_dir, pattern = paste0('^', output_file), full.names = TRUE)
   msg <- paste("Excel workbook created successfully! File located at", file_loc)
   message(msg)