+++ R/utilMWRyscale.R
@@ -1,8 +1,8 @@
 #' Get logical value for y axis scaling
 #'
 #' @param accdat \code{data.frame} for data quality objectives file for accuracy as returned by \code{\link{readMWRacc}}
-#' @param param character string to first filter results by a parameter in \code{"Characteristic Name"}
-#' @param yscl character indicating one of \code{"auto"} (default), \code{"log"}, or \code{"linear"}, see details
+#' @param param character string for the parameter to evaluate as provided in the \code{"Parameter"} column of \code{"accdat"}
+#' @param yscl character indicating one of \code{"auto"} (default), \code{"log"}, or \code{"linear"}
 #'
 #' @return A logical value indicating \code{TRUE} for log10-scale, \code{FALSE} for arithmetic (linear)
 #' @export
@@ -26,9 +26,9 @@
 #' 
 #' # log force
 #' utilMWRyscale(accdat, param = 'DO', yscl = 'log')
-utilMWRyscale <- function(accdat, param, yscl = c('auto', 'log', 'linear')){
+utilMWRyscale <- function(accdat, param, yscl = 'auto'){
   
-  yscl <- match.arg(yscl)
+  yscl <- match.arg(yscl, c('auto', 'log', 'linear'))
   
   # get scaling from accuracy data
   logscl <- accdat %>%